abstracted to organized components
